fastchg.c: disable forced fast charging
When enabled, this would force our device (in this case, Poco X3 NFC)
to charge as quickly as possible, even if the battery was at 45°C,
it wouldn't go down, sustaining ~5.7A or higher when charging with
the 33W charger that it came with.

After some testing, I found out that this was the culprit and by
setting the value to 0, we're effectively allowing the phone to
charge as intended and prevent it's battery from overheating.

Now, we fix it.