Cirrus CI: Try to get builds working again.
As the name of the main comment states, we'll try to fix
our Cirrus CI builds by getting the necessary packages installed
through our build script, if this works as expected, the functionality
of installing all the necessary packages will be moved over to another
script in the source, which will take care of all the packages and then
call cirrus_build.sh to then build the kernel.

It's not entirely necessary for us to fix Cirrus CI builds, however
it would be nice to have them working again and not have them show up
as a failed check on our sources when it's the image itself that doesn't
get the job done.