defconfig: enable CPU input boosting
I originally intended to make this only available on certain conditions.

Those conditions being our KProfiles
and whether they were set
to performance or not, but I found
that this would go against the boosting
behavior that we have up until this point.

Which leaves us with a bit of tuning that
we're probably going to need later.

The values we're taken from
https://github.com/Ardjlon/android_kernel_xiaomi_surya/blob/13/arch/arm64/configs/surya_defconfig#L831
We'll be using them for reference and tune them later.